Only themeprops in shared
