+++ web/src/nav/NavLinks.test.tsx
@@ -9,7 +9,7 @@ import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { KeyboardShortcutsProps } from '../../../shared/src/keyboardShortcuts'
 import { eventLogger } from '../tracking/eventLogger'
 import { NavLinks } from './NavLinks'
-import { ThemePreference } from '../../../shared/src/theme'
+import { ThemePreference } from '../search/theme'
 
 // Renders a human-readable list of the NavLinks' contents so that humans can more easily diff
 // snapshots to see what actually changed.
