+++ web/src/nav/NavLinks.tsx
@@ -16,11 +16,12 @@ import {
     KEYBOARD_SHORTCUT_SWITCH_THEME,
     KeyboardShortcutsProps,
 } from '../../../shared/src/keyboardShortcuts'
-import { ThemePreferenceProps, ThemeProps } from '../../../shared/src/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 import { EventLoggerProps } from '../tracking/eventLogger'
 import { fetchAllStatusMessages, StatusMessagesNavItem } from './StatusMessagesNavItem'
 import { UserNavItem } from './UserNavItem'
 import { CampaignsNavItem } from '../enterprise/campaigns/global/nav/CampaignsNavItem'
+import { ThemePreferenceProps } from '../search/theme'
 
 interface Props
     extends SettingsCascadeProps,
