+++ web/src/nav/UserNavItem.tsx
@@ -5,8 +5,9 @@ import { Link } from 'react-router-dom'
 import { ButtonDropdown, DropdownItem, DropdownMenu, DropdownToggle } from 'reactstrap'
 import * as GQL from '../../../shared/src/graphql/schema'
 import { KeyboardShortcut } from '../../../shared/src/keyboardShortcuts'
-import { ThemePreference, ThemePreferenceProps, ThemeProps } from '../../../shared/src/theme'
+import { ThemeProps } from '../../../shared/src/theme'
 import { UserAvatar } from '../user/UserAvatar'
+import { ThemePreferenceProps, ThemePreference } from '../search/theme'
 
 interface Props extends ThemeProps, ThemePreferenceProps {
     location: H.Location
